machine,uart: modify uart_recv to non-blocking
* the priv-1.10 kernel will poll the uart when timer intr comes

+++ machine/uart.h
@@ -10,7 +10,7 @@ extern void uart_init();
 extern void uart_send(uint8_t);
 extern void uart_send_string(const char *str);
 extern void uart_send_buf(const char *buf, const int32_t len);
-extern uint8_t uart_recv();
+extern int uart_recv();
 extern uint8_t uart_read_irq();
 extern uint8_t uart_check_read_irq();
 extern void uart_enable_read_irq();
